refactor: small changes to types and names

+++ samples/null_mapping/src/main.cpp
@@ -3,7 +3,7 @@
 #include <mmd/mmd.h>
 
 int main() {
-    int *val = (int *)0xffffffffc0287010;
+    unsigned long *val = (unsigned long *)0xffffffffc0287010;
     int err = mem_allocate(reinterpret_cast<unsigned long>(val), sizeof(int));
     if (err < 0) {
         std::cout << "Something went wrong: " << err << '\n';
